feat(sms command dataset forms): improve larg form performance
